v.1.1-change-circuit-pinouts
Made minor changes to Fig. 3 in the design document, updating the STM32 pinouts. Specifically, PA10 has been modified to function as an input pin. This change reflects an adjustment in the hardware configuration to accommodate the project's requirements. Developers and hardware integrators should take note of this alteration when referencing the pinout diagram.